requirements.txt
ffmpeg-python==0.2.0
openai-whisper==20230314
pyyaml==6.0
torch==2.2.0
